refactor: upgrade storybook to 6.1.18
